First step
I defined folders and files.
I placed my Phaser.min file inside my src folder.
Within my script.js file I required Phaser to start working with the data contained in said file (.min).
Also in my script.js I created the configuration variable where it contains an object, which determines the dimensions and characteristics of the game. I also defined the scene, this will call my three functions that I created outside of the configuration variable.
These three functions are: preload(), create() and update().
The first will be in charge of entering all the data that will be loaded before the game starts.
In the second what we are going to do is add the different objects. For example: platforms, background, etc.
And finally update. In this we are going to see the movements made by the player or user, since it is updated every second.